+++ CHANGELOG.rst
@@ -1,6 +1,10 @@
 Change Log
 ==========
 
+Unreleased
+----------
+* Feature: FileStore wrapper around provider files
+
 0.0.1 - Alpha
 -------------
 * Feature: Local File Storage